+++ Example/NavigationFromTab/NavigationFromTab/FirstView.swift
@@ -20,6 +20,7 @@ struct FirstView: View {
         VStack {
             Text("First")
             Button {
+                // True! It's can fix warning to console because .navigationStorageDestination is called outside of the TabView.
                 navigationStorage?.changeDestination(with: TabNavigation.second)
             } label: {
                 Text("to Second")