+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 
 All notable changes to this project will be documented in this file.
 
-## [1.11.0] - 2024-12-20
+## [1.11.0] - 2024-12-26
 
 #### Added